add Spotlight lift up code
